// types.ts
export interface FormData {
  name: string;
  email: string;
  phone: string;
  company: string;
  roles: string;
  agreed: boolean;
}

export interface FAQItem {
  question: string;
  answer: string;
}

export interface Testimonial {
  quote: string;
  author: string;
  designation: string;
  company: string;
}